+++ index.js
@@ -1,13 +1,12 @@
 // Packages
 const fs = require('fs');
 const { Client, Intents, Collection, MessageEmbed } = require('discord.js');
-const mcping = require('mcping-js');
 const { logger } = require('./logging.js');
+const { runTasks } = require('./tasks');
 
 // Config
-const { statusInterval, commandRefreshInterval, mcServerPort, version } = require('./config.json');
-const { token, mcServerIP } = require('./hidden.json');
-const deployCommands = require('./deploy-commands');
+const { version } = require('./config.json');
+const { token } = require('./hidden.json');
 
 // Creates the bot client
 const client = new Client({
@@ -33,90 +32,9 @@ for(const folder of commandFolders) {
 // Runs when the bot is online
 client.once('ready', async () => {
 	logger.child({ mode: 'STARTUP' }).info('Bot is initializing...');
-
-	await updateCommands();
-
+	await runTasks(client);
 	logger.child({ mode: 'STARTUP' }).info('Bot is active');
-
-	displayServer();
-	setInterval(displayServer, statusInterval * 1000);
 });
-
-// Refreshes the commands on startup
-const updateCommands = async () => {
-	const raw = fs.readFileSync('./storage.json');
-	const data = JSON.parse(raw);
-
-	const currentTime = Date.now();
-
-	// Refreshes commands only if it hasn't within the last 30 minutes
-	if(data.commandUpdate < currentTime) {
-		await deployCommands.execute(client);
-		data.commandUpdate = currentTime + (commandRefreshInterval * 60000);
-
-		// Updates time
-		fs.writeFileSync('./storage.json', JSON.stringify(data));
-	} else {
-		logger.child({ mode: 'STARTUP' }).info('Commands will be refreshed on startup in ' + Math.floor((data.commandUpdate - currentTime) / 60000) + ' minutes');
-	}
-};
-
-
-// Updates the bot's status periodically
-const displayServer = () => {
-	const server = new mcping.MinecraftServer(mcServerIP, mcServerPort);
-
-	let activity = '';
-	let status = 'dnd';
-
-	server.ping(1000, 757, (err, res) => {
-		// Server offline
-		if (!(typeof err === 'undefined' || err === null)) {
-
-			client.user.setPresence({
-				activities: [{
-					name: 'an offline server :(',
-					type: 'PLAYING',
-				}],
-				status: 'dnd',
-			});
-
-			// console.log(err);
-			return;
-		}
-
-		// Server online with no players
-		if (typeof res.players.sample === 'undefined') {
-			status = 'idle';
-			activity = res.players.online + '/' + res.players.max + ' players';
-		}
-
-		// Gets the online players
-		let onlinePlayers = [];
-
-		// Server online with players
-		if (!(typeof res.players.sample === 'undefined')) {
-			status = 'online';
-
-			for (let i = 0; i < res.players.sample.length; i++) {
-				onlinePlayers.push(res.players.sample[i].name);
-			}
-			onlinePlayers = onlinePlayers.sort().join(', ');
-
-			activity = res.players.online + '/' + res.players.max + ' players -\n ' + onlinePlayers;
-		}
-
-		// Sets the activity to the amount of players on the server
-		client.user.setPresence({
-			activities: [{
-				name: activity,
-				type: 'PLAYING',
-			}],
-			status: status,
-		});
-		logger.child({ mode: 'STATUS' }).debug(`Status has been updated with status '${status}' and activity '${activity}'`);
-	});
-};
 
 // Command Handeling
 client.on('interactionCreate', async interaction => {